refactor(kernel-handle): split role-trait god-file into per-domain modules
The 2017-line crates/librefang-kernel-handle/src/lib.rs held all 20 role
traits in one file. #3746 split the god-trait into role traits logically
but kept them co-located; this splits the file to match.

- One module file per capability domain: agent_control, memory_access,
  wiki_access, task_queue, event_bus, knowledge_graph, cron_control,
  approval_gate, hands_control, a2a_registry, channel_sender,
  prompt_store, workflow_runner, goal_control, tool_policy, api_auth,
  session_writer, catalog_query, plus acp_fs / acp_terminal (each holds
  its Bridge + Client trait and result type).
- lib.rs now keeps only the shared KernelOpError alias + KernelResult,
  the module decls/re-exports, the KernelHandle supertrait + blanket
  impl, and the prelude. Every `librefang_kernel_handle::*` path and the
  prelude resolve unchanged via `pub use <module>::*`.
- Each module imports shared types via `use super::*`; async_trait is
  imported only by modules defining async traits. mod/use blocks are
  alphabetical and imports grouped to match rustfmt + repo style.
- StubKernel object-safety test harness moved to tests.rs.
- Drift fix: the supertrait comment said "19 role traits"; it is 20
  (CatalogQuery added in #4842).

Pure structural change: no trait method, signature, default impl, or
public symbol altered; the crate's exported surface is identical.

Verification: static checks only (complete trait/struct redistribution,
brace balance, rustfmt-style import grouping + alphabetical mod/use
blocks, no trailing whitespace / double blanks / missing final newline).
No native cargo/rustfmt and no working Docker in this environment, so
cargo check/clippy/test/fmt were not run locally — relying on CI.
